consolidate main code
